move metadata code to esri.py
