+++ dcpy/lifecycle/package/shapefiles.py
@@ -6,8 +6,6 @@ from pathlib import Path
 
 import typer
 
-from dcpy.lifecycle import product_metadata
-from dcpy.models.data.shapefile_metadata import Attr, Edom
 from dcpy.models.product.dataset.metadata import (
     COLUMN_TYPES,
     ColumnValue,
@@ -15,9 +13,6 @@ from dcpy.models.product.dataset.metadata import (
     DatasetColumn,
     Metadata,
 )
-from dcpy.models.product.metadata import OrgMetadata
-from dcpy.utils.geospatial import esri_metadata, fgdb
-from dcpy.utils.geospatial.shapefile import Shapefile
 from dcpy.utils.logging import logger
 
 _shapefile_to_dcpy_types: dict[str, COLUMN_TYPES] = {
@@ -135,119 +130,3 @@ def _write_shapefile_metadata(
     out_path = output_path or Path("./metadata.yml")
     parse_shapefile_metadata(shp_xml_path).write_to_yaml(out_path)
     logger.info(f"Wrote metadata to {out_path}")
-
-
-@app.command("write_metadata")
-def _write_metadata(
-    product_name: str,
-    dataset_name: str,
-    path: Path,
-    layer: str,
-    org_md_path: Path | None = typer.Option(
-        None,
-        "--org-md-path",
-        help="Path to organizational metadata",
-    ),
-    zip_subdir: str | None = typer.Option(
-        None,
-        "--zip-subdir",
-        help="Directory structure within zip file, if relevant",
-    ),
-):
-    write_metadata(
-        product_name=product_name,
-        dataset_name=dataset_name,
-        path=path,
-        layer=layer,
-        zip_subdir=zip_subdir,
-        org_md=org_md_path,
-    )
-    logger.info(f"Wrote metadata to {layer} in {path}")
-
-
-def write_metadata(
-    product_name: str,
-    dataset_name: str,
-    path: Path,
-    layer: str,
-    zip_subdir: str | None,
-    org_md: Path | OrgMetadata | None,  # Allow passing OrgMetadata for testing purposes
-):
-    """Write product metadata to the shapefile metadata XML. Generates a new XML with defaults,
-    and applies additional product-specific values.
-
-    Args:
-        product_name (str): Name of product. e.g. "lion"
-        dataset_name (str): Name of dataset within a product. e.g. "pseudo-lots"
-        path (Path): Path to parent directory or zip file containing shapefile, or geodatabase.
-        layer (str): Shapefile or feature class name.
-        zip_subdir (str | None): Internal path if shp is nested within a zip file.
-            Must be None for file geodatabase.
-        org_md (Path | OrgMetadata | None): Metadata reference used to populate shapefile metadata.
-    """
-    if isinstance(org_md, Path) or not org_md:
-        org_md = product_metadata.load(org_md_path_override=org_md)
-
-    product_md = org_md.product(product_name).dataset(dataset_name)
-
-    metadata = esri_metadata.generate_metadata()
-
-    # Set dataset-level values
-    # TODO: define DCP organizationally required metadata fields
-    metadata.md_hr_lv_name = product_md.attributes.display_name
-    metadata.data_id_info.id_abs = product_md.attributes.description
-    metadata.data_id_info.other_keys.keyword = product_md.attributes.tags
-    metadata.data_id_info.search_keys.keyword = product_md.attributes.tags
-
-    metadata.eainfo.detailed.name = product_md.id
-    metadata.eainfo.detailed.enttyp.enttypl.value = product_md.id
-    metadata.eainfo.detailed.enttyp.enttypt.value = "Feature Class"
-
-    # Build attribute metadata for each column
-    metadata.eainfo.detailed.attr = [
-        _create_attr_metadata(column) for column in product_md.columns
-    ]
-
-    if ".gdb" in path.suffixes:
-        if zip_subdir is not None:
-            raise ValueError(
-                "Nested zipped GDBs are not supported. The GDB must be at the top level of the zip."
-            )
-        fgdb.write_metadata(gdb=path, layer=layer, metadata=metadata, overwrite=True)
-
-    elif ".shp" in path.suffixes or layer.endswith(".shp"):
-        shp = Shapefile(path=path, shp_name=layer, zip_subdir=zip_subdir)
-        shp.write_metadata(metadata, overwrite=True)
-
-
-def _create_attr_metadata(column: DatasetColumn) -> Attr:
-    """Create an Attr metadata object from a column specification."""
-    attr = Attr()
-
-    attr.attrlabl.value = "FID" if column.id == "uid" else column.id
-    attr.attalias.value = "FID" if column.name == "uid" else column.name
-    attr.attrdef.value = column.description
-
-    # TODO: define column-level defaults (e.g. attrdefs = 'Esri' if column.name == 'uid')
-    # TODO: map DCP types to Esri types (e.g. attrtype = 'OID' if column.name == 'uid') Note DCP types != Esri types
-    # attr.attrtype.value = column.data_type
-    # attr.attwidth.value = None
-    # attr.atprecis.value = None
-    # attr.attscale.value = None
-    # attr.attrdefs.value = ""
-
-    # Handle domain values if present
-    if hasattr(column, "values") and column.values:
-        attr.attrdomv.edom = [_create_edom_metadata(value) for value in column.values]
-
-    # TODO: handle 'attrdomv.udom', with other esri value defaults
-    return attr
-
-
-def _create_edom_metadata(column_value: ColumnValue) -> Edom:
-    """Create an Edom metadata object from a column value specification."""
-    edom = Edom()
-    edom.edomv = column_value.value
-    edom.edomvd = column_value.description
-
-    return edom
